backend/pyproject.toml
[tool.poetry]
name = "cms-automation-backend"
version = "0.1.0"
description = "AI-powered CMS automation backend using Claude Computer Use API"
authors = ["CMS Automation Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
celery = {extras = ["redis"], version = "^5.3.4"}
redis = "^5.0.1"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
structlog = "^23.2.0"
httpx = "^0.25.2"
pgvector = "^0.2.3"
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
asyncpg = "^0.30.0"
greenlet = "^3.2.4"
anthropic = ">=0.40.0"
nest-asyncio = "^1.6.0"
google-auth = "^2.24.0"
google-auth-oauthlib = "^1.1.0"
google-auth-httplib2 = "^0.1.1"
google-api-python-client = "^2.108.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
ruff = "^0.1.6"
mypy = "^1.7.1"
httpx = "^0.25.2"
faker = "^20.1.0"
safety = "^3.6.2"
bandit = "^1.8.6"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
